+++ evaluator/evaluator_test.go
@@ -4,6 +4,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/textwire/textwire/v2/fail"
 	"github.com/textwire/textwire/v2/lexer"
 	"github.com/textwire/textwire/v2/object"
 	"github.com/textwire/textwire/v2/parser"
@@ -419,5 +420,49 @@ func TestEvalComments(t *testing.T) {
 
 	for _, tc := range tests {
 		evaluationExpected(t, tc.inp, tc.expected)
+	}
+}
+
+func TestErrorHandling(t *testing.T) {
+	cases := []struct {
+		inp string
+		err *fail.Error
+	}{
+		{
+			inp: "{{ 5.somefunction() }}",
+			err: fail.New(
+				1,
+				"/path/to/file",
+				"evaluator",
+				fail.ErrNoFuncForThisType,
+				"somefunction",
+				object.INT_OBJ,
+			),
+		},
+		{
+			inp: "{{ 3 / 0 }}",
+			err: fail.New(
+				1,
+				"/path/to/file",
+				"evaluator",
+				fail.ErrDivisionByZero,
+			),
+		},
+	}
+
+	for _, tc := range cases {
+		evaluated := testEval(tc.inp)
+
+		errObj, ok := evaluated.(*object.Error)
+
+		if !ok {
+			t.Fatalf("evaluation failed: %s", errObj.String())
+		}
+
+		expect := tc.err.String()
+
+		if errObj.String() != expect {
+			t.Fatalf("error message is not '%s', got '%s'", expect, errObj.String())
+		}
 	}
 }