+++ evaluator/float_functions.go
@@ -1,9 +1,0 @@
-package evaluator
-
-import "github.com/textwire/textwire/v2/object"
-
-// floatIntFunc returns the integer part of the given float
-func floatIntFunc(receiver object.Object, args ...object.Object) object.Object {
-	floatVal := receiver.(*object.Float).Value
-	return &object.Int{Value: int64(floatVal)}
-}