+++ fail/fail.go
@@ -35,11 +35,18 @@ const (
 	ErrUnknownTypeForOperator  = "unknown type '%s' for '%s' operator"
 	ErrPrefixOperatorIsWrong   = "prefix operator '%s' cannot be applied to '%s'"
 	ErrUseStmtMustHaveProgram  = "the 'use' statement must have a program attached"
-	ErrNoFuncForThisType       = "function '%s' doesn't exist for type '%s'"
 	ErrLoopVariableIsReserved  = "the 'loop' variable is reserved. You cannot use it as a variable name"
 	ErrVariableTypeMismatch    = "cannot assign variable '%s' of type '%s' to type '%s'"
 	ErrDotOperatorNotSupported = "the dot operator is not supported for type '%s'"
 	ErrPropertyNotFound        = "property '%s' not found in type '%s'"
+	ErrDivisionByZero          = "division by zero error. The right-hand side of the division operator must not be zero"
+
+	// Functions
+	ErrNoFuncForThisType  = "function '%s' doesn't exist for type '%s'"
+	ErrFuncRequiresOneArg = "function '%s' on type '%s' requires at least one argument"
+	ErrFuncFirstArgInt    = "the first argument for function '%s' on type '%s' must be an integer"
+	ErrFuncSecondArgInt   = "the second argument for function '%s' on type '%s' must be an integer"
+	ErrFuncFirstArgStr    = "the first argument for function '%s' on type '%s' must be a string"
 
 	// Template errors
 	ErrUnsupportedType   = "unsupported type '%T'"
@@ -59,7 +66,7 @@ type Error struct {
 	message  string
 	line     uint
 	filepath string
-	origin   string // "parser" | "evaluator" | "template"
+	origin   string // "parser" | "evaluator" | "template" | "API"
 }
 
 // New creates a new Error instance of Error
@@ -80,8 +87,7 @@ func (e *Error) String() string {
 		path = fmt.Sprintf(" in %s", e.filepath)
 	}
 
-	return fmt.Sprintf("[Textwire ERROR%s:%d]: %s",
-		path, e.line, e.message)
+	return fmt.Sprintf("[Textwire ERROR%s:%d]: %s", path, e.line, e.message)
 }
 
 // FatalOnError calls log.Fatal if the error message is not empty